Fix for CMUCL bug in rtest12. Thanks to Ole Rohne, who offers the following reservations:
I don't particularly like this fix, because the use of
COMPILED-FUNCTION-P in KILL1 has nothing to do with a function being
compiled - on some systems, all functions are compiled. It is a kludge
to figure out if a function has been loaded from a lisp file and it
should be renamed accordingly. Being used exactly once in all of
maxima doesn't justify shadowing a symbol from COMMON-LISP