The file share/numeric/bffac.lisp does not contain lisp function definitions, it just contains maxima code in disguise, viz. the lisp code sets the MEXPR mproperties of various symbols like $BFFAC.  The original maxima code can easily be regenerated from this (except, of course, for the presumably copious comments ;-) and bffac.mac is the result of this automatic reconstruction.
The archive also contains bffac.bfz, which gives exactly the same
maxima code (the only difference between bffac.bfz and bffac.lisp is
that bffac.bfz correctly updates the $FUNCTIONS list, but, of course,
the generated mac file does this automatically).

See the comment at the beginning of bffac.mac for the details of the
conversion.