Put more info about (un)translated symbols in the UNLISP file.
transl.lisp ($declare_translated): Record the list of
$declare_translated symbols as value of the new global variable
*declared-translated-functions*.

(tr-mdefine-toplevel): Remove #+lispm from the piece of code which
gives symbols whose function definition is being translated the
ONCE-TRANSLATED property.

transs.lisp (translate-file): Bind *declared-translated-functions*.
(delete-with-side-effects-if): New auxiliary function.
(insert-necessary-function-declares): Rewrite.

Note that *untranslated-functions-called* can't contain doublets, so
the original loop can be simplified via delete-with-side-effects-if.

Previously, because of the #+lispm glitch described above, the only
source of symbols with a non-NIL ONCE-TRANSLATED property were those
given in $declare_translated.  So the users got just a subset of their
own input back, viz. those functions which were actually called.

The new behaviour is to write to the UNLISP file two lists: The
functions which are known to be translated without actually being in
the $declare_translated list, and those which are not known to be
translated.

The idea is that if everything seems to be OK you should get no output
at all from here.