Revised PART+ and PART*, which determine whether the operands of + and *, resp., match a simplification rule.
(1) Fixed bug which prevented operands of same type from matching.
(2) PART+ and PART* impute an argument equal to 0 or 1, resp., if
an argument fails to match, and 0 or 1 would match if it were present.
Fixed a bug in this code: the predicate which matched 0 or 1 was
not removed from the list (leading to a infinite loop).
(3) PART+ used to call $RATEXPAND and PART* use to call $FACTOR on the
input expression. As these can change the expression substantially,
and they are expensive operations in general, these have been disabled.
(4) Added code to the tellsimpafter match compiler to emit a call
to a function to print a message when a rule is triggered.
This code is emitted only if a global flag is enabled.

Tested as follows on cvs Maxima, as compiled with clisp and gcl.
I'm planning to collect these into one test file.

(1) run_testsuite -- succeeds.
(2) A test for some simplifications involving a`b (physical units) -- succeeds.
(3) A test for simplifications involving a`b (base 60 arithmetic) -- succeeds.
(4) Running examples (ineq.dem) of share/simplification/ineq.mac -- succeeds.
(5) Some additional examples -- succeeds.