A partial fix for Bug 938235: integrate((1/2)*u^2-1/u^5,u,1,sqrt(2)); is not correct..., and Bug 941457: integrate(1/x^5,x,1,2^(1/78)).
In function CV in src/defint.lisp, we apply the transformation using
symbolic values for the limits of integration.  Then substitute in the
actual limits.  This helps METHOD-BY-LIMITS see a "simpler" integrand
that $FACTOR can handle.

For both of these bugs, the failure is caused by maxima being unable
to factor the transformed result.  The above hack helps this, but more
work is clearly needed.