src/limit.lisp: Bug [ 1471861 ] limit(abs(sin(x)/x),x,0);
This fixes this bug.  In this case, MABS-SUBST was throwing to MABS
with a value of RETN.  If so, we try to compute the limit from both
sides and if that succeeds, we're done.

I think mabs-subst fails because it computes the limits abs(sin(x))
and abs(x) separately, and can't decide what to do with the 0/0 case.
But I'm not 100% sure on this.

src/rtest15.mac:
o Add a test for this bug