Make display, grind, and tex of ordinary floats respect fpprintprec.
Before, EXPLODEN (called by display code and grind) respected fpprec,
which isn't right because fpprec is a computational precision for bigfloats.
fpprintprec is the printing precision for bigfloats, so use that for
ordinary floats as well. TeX code didn't call EXPLODEN so it didn't
respect fpprec (nor fpprintprec).

This committed version has a couple of warts, both in EXPLODEN.
(1) EXPLODEN tries to print numbers in a nice way.
    The way it is now coded, the number of digits is often
    not exactly equal to the specified printing precision.
(2) EXPLODEN contains a literal constant 16. That is OK for
    64 bit floats but if ever there were 80 bit floats,
    some digits would be needlessly suppressed.

 - src/commac.lisp:
   in EXPLODEN, work with fpprintprec instead of fpprec

 - src/mactex.lisp:
   in TEXNUMFORMAT, call EXPLODEN instead of EXPLODE
   fix bug in MYPRINC which caused tex commands to be broken over end of line

 - doc/info/Floating.texi:
   rewrite description of fpprintprec

 - doc/info/descriptive.texi:
   work with fpprintprec instead of fpprec


Here is a test script. Output seems to be OK, accepting item (1) above as it stands.


x : 0.123456789012345689;
y : 0.123456789012345689b0;

float2bf : true;
linel : 200;

L : cons (0, makelist (i, i, 2, 20));

for fpprintprec in L do
    print ([x*1e20, x*1e10, x*1e5, x, x/1e5, x/1e10, x/1e20]);

for fpprintprec in L do
    print ([y*1e20, y*1e10, y*1e5, y, y/1e5, y/1e10, y/1e20]);

for fpprintprec in L do
    grind ([x*1e20, x*1e10, x*1e5, x, x/1e5, x/1e10, x/1e20]);

for fpprintprec in L do
    grind ([y*1e20, y*1e10, y*1e5, y, y/1e5, y/1e10, y/1e20]);

with_stdout ("/tmp/float-print-prec.tex",
    print ("\\documentclass{article}"),
    print ("\\begin{document}"),
    for fpprintprec in L do
        tex ([x*1e20, x*1e10, x*1e5, x, x/1e5, x/1e10, x/1e20]),
    for fpprintprec in L do
        tex ([y*1e20, y*1e10, y*1e5, y, y/1e5, y/1e10, y/1e20]),
    print ("\\end{document}"));