Bug [ 1073338 ] integrate yields incorrect result on rational function
This is a workaround for this bug.  We basically simplify some of the
intermediate results so that maxima doesn't get confused.  I think
this is a gcd problem.

src/residu.lisp:
o Fix a few comments and add a few more.
o For the case of simple poles, use RES1 instead of the more general
  $RESIDUE.  (Workaround 1)
o In RES1, call $RECTFORM on each pole to simplify somewhat so that
  maxima doesn't get confused.  (Workaround 2)

tests/rtest15.mac:
o Add a test for 1073338, but disable it for now.  The integral takes
  a bit of time to compute, and the result is quite messy.  If
  sqrtdenest were available, the result can be simplified quite a bit.