Bug [ 1469411 ] limit(t^2*exp(-4*t/3-8*exp(-t)),t,inf) is wrong Bug [ 1594330 ] limit(x*(atan(x)-%pi/2),x,inf)
These are bugs in the L'Hospital heuristic.  Maxima was sometimes
converting the form n/d to (1/d)/(1/n) which actually makes the
problem harder.

src/limit.lisp (lhop-denom):
o If the numerator has an exponential and the denominator is a
  polynomial, don't invert them.
o If the denominator has an exponential and the numerator is a
  polynomial, don't invert them.
o If the numerator involves asin, acos, or atan, and the denominator
  or the reciprocal of the denominator is a polynomial, don't invert
  them.


tests/rtest16.mac:
o Add tests for 1469411 and 1594330.