Revise documentation for describe and for ? shortcut to say that exact search is the default. I have revised all of the existing documentation files and I believe I have done OK (in part because the relevant text is not translated in all languages yet), except for German, which definitely needs to be reviewed. Other languages can also benefit from review.
In addition, I have cut out the @node "Introduction to Help" from all languages,
because it was only a duplicate of text which also appears in "describe".

With these changes, make succeeds for es and es-utf8, but not any variety of pt.
I believe the make problems with pt are due to some other problem.
Other languages are not in the make system yet.