Bug [ 1714044 ] Maxima asks unnecessary questions in integration
Make EASY-SUBS handle a few more cases.  In particular, if the result
doesn't involve inverse functions and if the denominator doesn't
involve the integration variable, we can try to substitute the limits
in directly.

This change gets rid of the questions which are produced when limit is
used to substitute in the limits.