Bug [1161505] specint returns Lisp symbol ...
Add code in the routine NEGTEST to set the flag *hyp-return-noun-form-flag* which signals that a noun form has to be returned.

Bug [1376860] specint(gammaincomplete ... seems wrong

Change the transformation of the Gammaincomplete(a,x) function to a Gammgreek function for a<>0. Only for a=0 we transform to a Whittaker W function. Now we get more simple and the expected results.

The name of the function to transform the Gammaincomplete function  is changed to gammaincomplete-to-gammagreek. The old routine with the name gammincompletetw is still present, but no longer called.

Dieter Kaiser