Fix some compiler warnings/notes from clisp.  These are true issues.
clmacs.lisp:
o Ignore the low and high arguments for the flonum deftype.

macsys.lisp:
o Declare S ignorable in $SYSTEM.  Not all Lisp's use this.

numerical/slatec/quadpack.lisp:
o The condition object E is not used, so remove it.
o It appears as if *PLOT-REALPART* is not known to be a special
  variable when this is compiled, and clisp warns about this variable
  not being used.  Declare the variable as special.  (May the order of
  compilation should be changed?)