Changing the Zeta function to a simplifying function
A simplifying function simp-gamma has been implemented. The noun and verb properties are put on the property list.

The algorithm is almost the same as the algorithm of the old $zeta function. The direct calls to internal functions of the simplifier have been removed. The handling of the specal cases for the numbers 0 and 1 have been complemented. A call to resimplify has been removed.

Tested with GCL 2.6.8 and CLISP 1.44. The testsuite has no problems.