Adding the Laplace transform of expintegral_e and expintegral_e1.
Correcting a bug in gamma_incomplete-to-gammagreek:
For a negative order a of the incomplete gamma function the transformation to gammagreek is not possible. The gamma function in the transformation is not defined for negative integers.

Two more tests in rtest_hypgeo.mac give a result.

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.