Improving the handling of complex expressions for the functions abs, carg and rectform, ...:
Remove the call to rectform in $cabs. We get much simpler results for $cabs.

Do not use %cabs as a noun form but mabs. This is more consistent and allows further  simplifications.

Do not use asksign but $sign to determine the sign of a symbol. When the sign is unknown return a  more general form. With this change the algorithm do not stop when simplifying expressions.

Extend absarg-mabs to test if we have a complex expression. This test avoids an additional  simplification wich will cause an endless loop.