Updating testsuite files because of changes of the functions for the complex components:
rtest_hypgeo.mac:
We no longer get a noun form with cabs but with abs for two examples.

rtest_integrate.mac:
Do not use a complex symbol but a real symbol for integrals wich involve the function sqrt.

rtest_trig.mac:
We no longer get a noun form with cabs but with abs for two examples.

testsuite.lisp:
The examples 42 and 43 in rtest_abs.mac give the expected results.