Adding a function learn-abs.
When a pattern with abs(x)< a, a a positive expression, is found, the facts x<a and -x<a are learned too. This works for assume(abs(x)<a) and forget(abs(x)<a) symmetrically.

The function learn-abs uses the modified functionality of the function isinop in simp.lisp.

Related bug report:
SF [1041570] assume(abs(x)<1) should imply x<1 and x>-1

Some examples are added to the testfile rtest_sign.mac.

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.