Merged in the following revision to the branch 5.19:
compar.lisp 1.52 -> 1.50

Changes in revision 1.52:

When in $csign set limitp to NIL to avoid endless loop between trisplit - sratsimp - sign-prep.
Related bug report: SIGN-PREP strangeness - ID: 2835098
Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.

Changes in revision 1.51:

o change pred-reverse to call simp-mnot (via take);  fix for bug 1725951
o insert some calls to simplify in simp-mnot; fix for bug 2836339


rpart.lisp 1.27 ->1.26

Changes in revision 1.27:

Return a real answer for a subsripted variable which is not declared to be complex.
integrate(f(t),t,x[1],x[2]) works again.
Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.

defint 1.65 -> 1.64:

Set limitp not in the routine $defint but in the routine defint.

The answer of $imagpart depends on the value of limitp. In the routine $defint we prepare the integrand and check the boundaries of the integral using the routine $imagpart. To get  the expected answer from $imagpart we do not set limitp in $defint, but later in the routine defint.

Related bug report: defint fails to determine if one of its limit is real - ID: 2840566

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.