Adding a routine learn-numer.
learn-numer evaluates constant expressions numerically and adds the result to the assume database. The facts are put into the context '$learndata, if the current context is the user context '$initial. Otherwise the facts are put into the current context.

In the routine killcontext look, if the context is still on the list of active contexts. Remove it from the list, before deleting.

Related bug report:
Bug ID: 2477795 - "assume":problems with fractions or multiples of %pi and %e

No problems with the testsuite and the share_testsuite.