Implementing a general mechanism for functions to map over bags.
The implementation works for an arbitrary number of arguments and for an arbitrary operator.

The functionality can be switched off with the flag $distribute_over. The standard value is T.

A user access to this functionality has been not added at this time.

No problems with the testsuite and share_testsuite.