Correcting the routine dionarghyp:
It is wrong to apply the algorithm for the Laplace transform of a hypergeometric function to an argument with a negative power, e.g. for hypergeometric([a, ...], [b, ...], t^-1).

We check that the power of the argument is a positive integer.

Changes in rtest_hypgeo.mac:
Example 226 now gives a correct noun form. Two examples for the function asin and atan now gives a noun form too.

No problems with the testsuite and share_testsuite.