In solvequartic do not call simpnrt, but power to calculate d and e. Furthermore, use the functions add, sub and div to do the calculations.
It is possible to simplify the algorithm further and to cut out all direct calls of simplifying functions and some unnecessary calls of the simplifier. The change of this revision solves the problem that simpnrt hangs, when called directly with an expression which contains parameters declared to be constant.

Related bug report:
2036462 - Very long calculation time, normal ?

No problems with the testsuite. In the file rtest_to_poly_solver the ordering of the results of one example has changed.