Implementing a more simple formula which corresponds to the definition of acosh for the logarc of acosh. The formula we have is optimized for numerical evaluation, but is more complicated for symbolic evaluation.
This affects the logarc of the asech function too. Again the result of lorgarc corresponds to the definition of the asech function.

Avoiding the use of the global special variable 1//2 for the logarc of atan2. Calling the function root for this case.

Related bug report:
ID: 3040667 - logarc(acosh(z)) needlessly complicated