ratmac.lisp:
Cutting out the definition of $ratvarswitch from the file ratmac.lisp.
The option variable $ratvarswitch is only used in suprv1.lisp.

suprv1.lisp:

Adding the definition of $ratvarswitch to the file suprv1.lisp.
Define $ratvarswitch with defmvar, so it is resetable to the default value.

Cutting out $ratvarswitch from the definition declare-top.

lmdcls.lisp:

Adding the option variable $ratvarswitch.

No problems with the testsuite and share_testsuite.