adds to list of exported symbols adds gcl-specific eval-whens added tracer to pregexp-recur adds a special list of meta-chars (buried in pregexp-quote before)
big change:
we implement single/multi-line mode by inserting markers in regexp
parse tree and then, after the parse tree's creation, we filter it
and modify :any/:all, :eos/:eol :bos/:bol accordingly.

we use two new functions (pregexp-bol / -eol) and pregexp-filter-line-modes
to do this.

small change:
I believe that lookbehind/ahead is incorrectly implemented.
In any event, :any is not the correct char class to be backing
up on/advancing on, so I made these :all classes.

pregexp-quote:
This was an infinite loop, as counter was not decremented.
Added a test, too.